+++ tests/wl_tests_measures/test_measures_adjusted_freq.py
@@ -17,20 +17,30 @@
 # ----------------------------------------------------------------------
 
 from tests import wl_test_init
+from tests.wl_tests_measures import test_measures_dispersion
 from wordless.wl_measures import wl_measures_adjusted_freq
 
 main = wl_test_init.Wl_Test_Main()
 
 def test_to_freq_sections_items():
-    items_search = ['w1', 'w2']
-    items = ['w1'] * 7 + ['w2'] * 3
+    assert wl_measures_adjusted_freq.to_freq_sections_items(
+        main,
+        test_measures_dispersion.ITEMS_TO_SEARCH,
+        test_measures_dispersion.ITEMS
+    ) == test_measures_dispersion.FREQ_SECTIONS_ITEMS
 
-    freq_sections_items = {
-        'w1': [2, 2, 2, 1, 0],
-        'w2': [0, 0, 0, 1, 2]
-    }
+# Reference: Gries, S. T. (2008). Dispersions and adjusted frequencies in corpora. International Journal of Corpus Linguistics, 13(4), 403–437. https://doi.org/10.1075/ijcl.13.4.02gri (p. 410)
+def test_fald():
+    assert round(wl_measures_adjusted_freq.fald(main, test_measures_dispersion.TOKENS, 'a'), 3) == 11.764
+    assert wl_measures_adjusted_freq.fald(main, test_measures_dispersion.TOKENS, 'aa') == 0
 
-    assert wl_measures_adjusted_freq.to_freq_sections_items(main, items_search, items) == freq_sections_items
+def test_farf():
+    assert round(wl_measures_adjusted_freq.farf(main, test_measures_dispersion.TOKENS, 'a'), 1) == 10.8
+    assert wl_measures_adjusted_freq.farf(main, test_measures_dispersion.TOKENS, 'aa') == 0
+
+def test_fawt():
+    assert round(wl_measures_adjusted_freq.fawt(main, test_measures_dispersion.TOKENS, 'a'), 3) == 9.328
+    assert wl_measures_adjusted_freq.fawt(main, test_measures_dispersion.TOKENS, 'aa') == 0
 
 # References:
 #     Carroll, J. B. (1970). An alternative to Juilland’s usage coefficient for lexical frequencies and a proposal for a standard frequency index. Computer Studies in the Humanities and Verbal Behaviour, 3(2), 61–65. https://doi.org/10.1002/
@@ -40,7 +50,6 @@ def test_carrolls_um():
     assert round(wl_measures_adjusted_freq.carrolls_um(main, [2, 1, 1, 1, 0]), 2) == 4.31
     assert round(wl_measures_adjusted_freq.carrolls_um(main, [4, 2, 1, 1, 0]), 3) == 6.424
     assert round(wl_measures_adjusted_freq.carrolls_um(main, [1, 2, 3, 4, 5]), 3) == 14.108
-
     assert wl_measures_adjusted_freq.carrolls_um(main, [0, 0, 0, 0, 0]) == 0
 
 # References
@@ -51,7 +60,6 @@ def test_juillands_u():
     assert round(wl_measures_adjusted_freq.juillands_u(main, [0, 4, 3, 2, 1]), 2) == 6.46
     assert round(wl_measures_adjusted_freq.juillands_u(main, [2, 2, 2, 2, 2]), 0) == 10
     assert round(wl_measures_adjusted_freq.juillands_u(main, [4, 2, 1, 1, 0]), 3) == 4.609
-
     assert wl_measures_adjusted_freq.juillands_u(main, [0, 0, 0, 0, 0]) == 0
 
 # References:
@@ -62,7 +70,6 @@ def test_rosengres_kf():
     assert round(wl_measures_adjusted_freq.rosengrens_kf(main, [2, 2, 2, 2, 1]), 2) == 8.86
     assert round(wl_measures_adjusted_freq.rosengrens_kf(main, [4, 2, 1, 1, 0]), 3) == 5.863
     assert round(wl_measures_adjusted_freq.rosengrens_kf(main, [1, 2, 3, 4, 5]), 3) == 14.053
-
     assert wl_measures_adjusted_freq.rosengrens_kf(main, [0, 0, 0, 0, 0]) == 0
 
 # References:
@@ -71,17 +78,19 @@ def test_rosengres_kf():
 def test_engwalls_fm():
     assert round(wl_measures_adjusted_freq.engwalls_fm(main, [4, 2, 1, 1, 0]), 1) == 6.4
     assert round(wl_measures_adjusted_freq.engwalls_fm(main, [1, 2, 3, 4, 5]), 0) == 15
-
     assert wl_measures_adjusted_freq.engwalls_fm(main, [0, 0, 0, 0, 0]) == 0
 
 # Reference: Gries, S. T. (2008). Dispersions and adjusted frequencies in corpora. International Journal of Corpus Linguistics, 13(4), 403–437. https://doi.org/10.1075/ijcl.13.4.02gri (p. 409)
 def test_kromers_ur():
     assert round(wl_measures_adjusted_freq.kromers_ur(main, [2, 1, 1, 1, 0]), 1) == 4.5
-
     assert wl_measures_adjusted_freq.kromers_ur(main, [0, 0, 0, 0, 0]) == 0
 
 if __name__ == '__main__':
     test_to_freq_sections_items()
+
+    test_fald()
+    test_farf()
+    test_fawt()
 
     test_carrolls_um()
     test_juillands_u()