+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ## [3.1.0](https://github.com/BLKSerene/Wordless/releases/tag/3.1.0) - ??/??/2022
 ### 🎉 New Features
 - Settings: Add Settings - Measures - Dispersion - Gries's DP
+- Work Area: Add Wordlist/N-gram Generator - Generation Settings - Measures of Dispersion / Measure of Adjusted Frequency - Average Logarithmic Distance / Average Reduced Frequency / Average Waiting Time
 
 ### ✨ Improvements
 - File Area: Disallow empty file names